add support for multiple images on kitty
