+++ CMakeLists.txt
@@ -226,6 +226,7 @@ list(APPEND UEBERZUG_SOURCES
     "src/canvas/chafa.cpp"
     "src/canvas/kitty/kitty.cpp"
     "src/canvas/kitty/chunk.cpp"
+    "src/canvas/kitty/window/stdout.cpp"
     "src/canvas/iterm2/iterm2.cpp"
     "src/canvas/iterm2/chunk.cpp"
